Reorganised test classes and added placeholder stress tests
