+++ .gitignore
@@ -5,3 +5,4 @@
 composer.lock
 composer.phar
 phpunit.xml
+.phpunit.result.cache
