+++ phpunit.xml.dist
@@ -20,6 +20,9 @@
         <testsuite name="Rinvex Attributes Feature Test Suite">
             <directory suffix="Test.php">./tests/Feature</directory>
         </testsuite>
+        <testsuite name="Rinvex Attributes Stress Test Suite">
+            <directory suffix="Test.php">./tests/Stress</directory>
+        </testsuite>
     </testsuites>
     <filter>
         <whitelist>
