+++ tests/Feature/AttributeCreationTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Rinvex\Attributes\Tests\Feature;
 
 use Rinvex\Attributes\Tests\Stubs\User;
+use Rinvex\Attributes\Tests\TestCase;
 
 class AttributeCreationTest extends TestCase
 {
