+++ tests/Stress/AttributeTest.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rinvex\Attributes\Tests\Stress;
+
+use Rinvex\Attributes\Tests\Stubs\User;
+use Rinvex\Attributes\Tests\TestCase;
+
+class AttributeTest extends TestCase
+{
+    
+}
